Update show.html.erb
You should do  <% if current_user && @link.user == current_user %> instead of <% if @link.user == current_user %>
Because: When a user posts a link, then deletes his account the user_id of the link is "nil". Keep in mind that the id of the current_user is always nil when not logged in. So every "guest" will see the Edit and Destroy button then.